Add CPI test in program-test
