+++ program-test/tests/fuzz.rs
@@ -1,14 +1,8 @@
 use {
     solana_banks_client::BanksClient,
     solana_program::{
-        account_info::{next_account_info, AccountInfo},
-        entrypoint::ProgramResult,
-        hash::Hash,
-        instruction::{AccountMeta, Instruction},
-        msg,
-        program::invoke,
-        pubkey::Pubkey,
-        rent::Rent,
+        account_info::AccountInfo, entrypoint::ProgramResult, hash::Hash, instruction::Instruction,
+        msg, pubkey::Pubkey, rent::Rent,
     },
     solana_program_test::{processor, ProgramTest},
     solana_sdk::{
@@ -16,51 +10,26 @@ use {
     },
 };
 
-// Process instruction to invoke into another program
-fn invoker_process_instruction(
-    _program_id: &Pubkey,
-    accounts: &[AccountInfo],
-    _input: &[u8],
-) -> ProgramResult {
-    // if we can call `msg!` successfully, then InvokeContext exists as required
-    msg!("Processing invoker instruction before CPI");
-    let account_info_iter = &mut accounts.iter();
-    let invoked_program_info = next_account_info(account_info_iter)?;
-    invoke(
-        &Instruction::new(*invoked_program_info.key, &[0], vec![]),
-        &[invoked_program_info.clone()],
-    )?;
-    msg!("Processing invoker instruction after CPI");
-    Ok(())
-}
-
-// Process instruction to be invoked by another program
 #[allow(clippy::unnecessary_wraps)]
-fn invoked_process_instruction(
+fn process_instruction(
     _program_id: &Pubkey,
     _accounts: &[AccountInfo],
     _input: &[u8],
 ) -> ProgramResult {
     // if we can call `msg!` successfully, then InvokeContext exists as required
-    msg!("Processing invoked instruction");
+    msg!("Processing instruction");
     Ok(())
 }
 
 #[test]
 fn simulate_fuzz() {
     let rt = tokio::runtime::Runtime::new().unwrap();
-    let invoker_program_id = Pubkey::new_unique();
+    let program_id = Pubkey::new_unique();
     // Initialize and start the test network
-    let mut program_test = ProgramTest::new(
-        "program-test-fuzz-invoker",
-        invoker_program_id,
-        processor!(invoker_process_instruction),
-    );
-    let invoked_program_id = Pubkey::new_unique();
-    program_test.add_program(
-        "program-test-fuzz-invoked",
-        invoked_program_id,
-        processor!(invoked_process_instruction),
+    let program_test = ProgramTest::new(
+        "program-test-fuzz",
+        program_id,
+        processor!(process_instruction),
     );
 
     let (mut banks_client, payer, last_blockhash) =
@@ -74,8 +43,7 @@ fn simulate_fuzz() {
             &mut banks_client,
             &payer,
             last_blockhash,
-            &invoker_program_id,
-            &invoked_program_id,
+            &program_id,
         )
         .await
     });
@@ -84,18 +52,12 @@ fn simulate_fuzz() {
 #[test]
 fn simulate_fuzz_with_context() {
     let rt = tokio::runtime::Runtime::new().unwrap();
-    let invoker_program_id = Pubkey::new_unique();
+    let program_id = Pubkey::new_unique();
     // Initialize and start the test network
-    let mut program_test = ProgramTest::new(
-        "program-test-fuzz-invoker",
-        invoker_program_id,
-        processor!(invoker_process_instruction),
-    );
-    let invoked_program_id = Pubkey::new_unique();
-    program_test.add_program(
-        "program-test-fuzz-invoked",
-        invoked_program_id,
-        processor!(invoked_process_instruction),
+    let program_test = ProgramTest::new(
+        "program-test-fuzz",
+        program_id,
+        processor!(process_instruction),
     );
 
     let mut test_state = rt.block_on(async { program_test.start_with_context().await });
@@ -108,8 +70,7 @@ fn simulate_fuzz_with_context() {
             &mut test_state.banks_client,
             &test_state.payer,
             test_state.last_blockhash,
-            &invoker_program_id,
-            &invoked_program_id,
+            &program_id,
         )
         .await
     });
@@ -120,8 +81,7 @@ async fn run_fuzz_instructions(
     banks_client: &mut BanksClient,
     payer: &Keypair,
     last_blockhash: Hash,
-    invoker_program_id: &Pubkey,
-    invoked_program_id: &Pubkey,
+    program_id: &Pubkey,
 ) {
     let mut instructions = vec![];
     let mut signer_keypairs = vec![];
@@ -132,14 +92,10 @@ async fn run_fuzz_instructions(
             &keypair.pubkey(),
             Rent::default().minimum_balance(i as usize),
             i as u64,
-            invoker_program_id,
+            program_id,
         );
         instructions.push(instruction);
-        instructions.push(Instruction::new(
-            *invoker_program_id,
-            &[0],
-            vec![AccountMeta::new_readonly(*invoked_program_id, false)],
-        ));
+        instructions.push(Instruction::new(*program_id, &[0], vec![]));
         signer_keypairs.push(keypair);
     }
     // Process transaction on test network
